+++ CHANGELOG.md
@@ -11,6 +11,7 @@
   ```
 - Input: removed `compact` and `renderUnderline` props
 - Select: removed `Type.MATERIAL`
+- Toggle: added `Size.Size14` which is the new default
 - Components no longer accept `theme` prop, themes are managed using CSS Custom Properties instead. To apply a theme to your app or some part of it, wrap it into `ThemeProvider`:
   ```js
   import Theme, {ThemeProvider} from '@jetbrains/ring-ui/components/global/theme';